+++ build_resolvers/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 2.5.2-wip
+
 ## 2.5.1
 
 - Use `build_runner_core` 2.5.1.